Add files via upload
Diabetes prediction based on diagnostic measures

File for prediction of diabetes based on diagnostic measures

About DATA:
This dataset is originally from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective of the dataset is to diagnostically predict whether or not a patient has diabetes, based on certain diagnostic measurements included in the dataset. Several constraints were placed on the selection of these instances from a larger database. In particular, all patients here are females at least 21 years old of Pima Indian heritage.

The datasets consists of several medical predictor variables and one target variable, Outcome. Predictor variables includes the number of pregnancies the patient has had, their BMI, insulin level, age, and so on.

Columns
1.Pregnancies = Number of times pregnant

2.Glucose = Plasma glucose concentration a 2 hours in an oral glucose tolerance test

3.BloodPressure = Diastolic blood pressure (mm Hg)

4.SkinThickness = Triceps skin fold thickness (mm)

5.Insulin = 2-Hour serum insulin (mu U/ml)

6.BMI = Body mass index (weight in kg/(height in m)^2)

7.DiabetesPedigreeFunction = Diabetes pedigree function

8.Age = Age (years)

9.Outcome = Class variable (0 or 1) 268 of 768 are 1, the others are 0